Move module to free folder
