+++ front/app/modules/free/native_surveys/admin/components/SurveyBuilderToolbox/ToolboxItem.tsx
