+++ front/app/modules/free/native_surveys/admin/components/SurveyBuilderTopBar/index.tsx
