+++ front/app/modules/free/native_surveys/admin/components/Tab.tsx
