+++ front/app/modules/free/native_surveys/components/NativeSurveyFooter/index.tsx
