+++ front/app/modules/index.ts
@@ -4,7 +4,7 @@ import projectFoldersConfiguration from './commercial/project_folders';
 import smartGroupsConfiguration from './commercial/smart_groups';
 import userCustomFieldsConfiguration from './commercial/user_custom_fields';
 import ideaCustomFieldsConfiguration from './commercial/idea_custom_fields';
-import nativeSurveyConfiguration from './commercial/native_surveys';
+import nativeSurveyConfiguration from './free/native_surveys';
 import granularPermissionsConfiguration from './commercial/granular_permissions';
 import projectManagementConfiguration from './commercial/project_management';
 import ideaAssignmentConfiguration from './commercial/idea_assignment';
